fix(core)!: changed the order expected from `open`
This is inline with documentation and the way `diff` expects it's return
values.

This is a breaking change, however, I think it will mostly affect hooks,
not the core functionality. Namely this will change the value passed for
`winnr` to `post_open` and may cause `block_end` to not be called.

This also only people who were using `open` as a function, rather than a
string, and have already swapped the two values.